refactor: move the root stack navigator out of the main app file
